+++ react/__tests__/ProductContextProvider.test.tsx
@@ -1,20 +1,30 @@
 /* eslint-env jest */
 import React, { useContext } from 'react'
 import { render } from '@vtex/test-tools/react'
+
+// eslint-disable-next-line jest/no-mocks-import
 import { getProduct, getItem } from '../__mocks__/productMock'
 import ProductContextProvider from '../ProductContextProvider'
 import ProductContext from '../ProductContext'
 import ProductDispatchContext from '../ProductDispatchContext'
+
 const { useProductDispatch } = ProductDispatchContext
 
 const ProductPageMock = () => {
-  const { selectedItem, product, selectedQuantity } = useContext(ProductContext) as any
+  const { selectedItem, product, selectedQuantity } = useContext(
+    ProductContext
+  ) as any
+
   return (
     <div>
       <div>Product Page</div>
-      <div>Selected Item id: {selectedItem && selectedItem.itemId}</div>
-      <div>Selected Item name: {selectedItem && selectedItem.name}</div>
-      {product ? <div>product slug: {product && product.linkText}</div> : <div>no product</div>}
+      <div>Selected Item id: {selectedItem?.itemId}</div>
+      <div>Selected Item name: {selectedItem?.name}</div>
+      {product ? (
+        <div>product slug: {product?.linkText}</div>
+      ) : (
+        <div>no product</div>
+      )}
       <div>Selected Quantity: {selectedQuantity}</div>
     </div>
   )
@@ -28,6 +38,7 @@ describe('ProductContextProvider component', () => {
       query: {},
       ...customProps,
     }
+
     return props
   }
 
@@ -43,28 +54,37 @@ describe('ProductContextProvider component', () => {
     return {
       ...component,
       testNoProduct: () => getByText('no product'),
-      getSelectedItemId: (item: { itemId: string }) => getByText(`Selected Item id: ${item.itemId}`),
+      getSelectedItemId: (item: { itemId: string }) =>
+        getByText(`Selected Item id: ${item.itemId}`),
       getSelectedItemName: (item: { name: string }) =>
         getByText(`Selected Item name: ${item.name}`),
-      getProductSlug: (product: { linkText: string }) => getByText(`product slug: ${product.linkText}`),
-      rerender: (newProps: any) => rerender(<ProductContextProvider {...newProps}>
-        <ProductPageMock />
-      </ProductContextProvider>)
+      getProductSlug: (product: { linkText: string }) =>
+        getByText(`product slug: ${product.linkText}`),
+      rerender: (newProps: any) =>
+        rerender(
+          <ProductContextProvider {...newProps}>
+            <ProductPageMock />
+          </ProductContextProvider>
+        ),
     }
   }
 
   it('should render with product', () => {
     const product = getProduct()
     const { getProductSlug } = renderComponent()
+
     getProductSlug(product)
   })
 
   it('should render with no product and then switch', () => {
-    const { getProductSlug, rerender, testNoProduct } = renderComponent({ product: undefined })
+    const { getProductSlug, rerender, testNoProduct } = renderComponent({
+      product: undefined,
+    })
 
     testNoProduct()
 
     const product = getProduct()
+
     rerender({ product, query: {} })
 
     getProductSlug(product)
@@ -106,6 +126,7 @@ describe('ProductContextProvider component', () => {
     const newProduct = getProduct({
       items: [noQuantity, noQuantity, itemWithQuantity, otherItemWithQuantity],
     })
+
     const { getSelectedItemId, getSelectedItemName } = renderComponent({
       product: newProduct,
     })
@@ -121,11 +142,13 @@ describe('ProductContextProvider component', () => {
     const newProduct = getProduct({
       items: [itemone, itemtwo, itemthree],
     })
+
     const props = {
       product: newProduct,
       params: { slug: newProduct.linkText },
       productQuery: { product: newProduct, loading: false },
     }
+
     const {
       getSelectedItemId,
       getSelectedItemName,
@@ -144,12 +167,17 @@ describe('ProductContextProvider component', () => {
   it('should dispatch action with bad args and not break anything', () => {
     const BadComponent = () => {
       const dispatch = useProductDispatch()
-      dispatch && dispatch({ type: 'SET_QUANTITY', quantity: 1 } as any)
-      dispatch && dispatch({ type: 'SKU_SELECTOR_SET_VARIATIONS_SELECTED', quantity: 1 } as any)
-      dispatch && dispatch({ type: 'SKU_SELECTOR_SET_IS_VISIBLE', quantity: 1 } as any)
-      dispatch && dispatch({ type: 'SET_SELECTED_ITEM', quantity: 1 } as any)
-      dispatch && dispatch({ type: 'SET_ASSEMBLY_OPTIONS', quantity: 1 } as any)
-      dispatch && dispatch({ type: 'SET_PRODUCT', quantity: 1 } as any)
+
+      dispatch?.({ type: 'SET_QUANTITY', quantity: 1 } as any)
+      dispatch?.({
+        type: 'SKU_SELECTOR_SET_VARIATIONS_SELECTED',
+        quantity: 1,
+      } as any)
+      dispatch?.({ type: 'SKU_SELECTOR_SET_IS_VISIBLE', quantity: 1 } as any)
+      dispatch?.({ type: 'SET_SELECTED_ITEM', quantity: 1 } as any)
+      dispatch?.({ type: 'SET_ASSEMBLY_OPTIONS', quantity: 1 } as any)
+      dispatch?.({ type: 'SET_PRODUCT', quantity: 1 } as any)
+
       return <div>Bad Component rendered OK</div>
     }