+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- When variants are passed in the query string, guarantee that the matched SKU is available if possible.
 
 ## [0.8.0] - 2020-07-10
 
@@ -21,7 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.6.0] - 2019-12-27
 ### Added
-- Docs buidler.
+- Docs builder.
 
 ## [0.5.0] - 2019-11-11
 ### Added